+++ MAM_Assessment/Entertainment.cs
@@ -3,7 +3,7 @@ namespace MAM_Assessment
 {
     public class Entertainment : Product
     {
-        public Entertainment(double price, string description, int quantity) : base(price, description, quantity)
+        public Entertainment(decimal price, string description, int quantity) : base(price, description, quantity)
         {
 
         }
