+++ MAM_Assessment/Food.cs
@@ -3,7 +3,7 @@ namespace MAM_Assessment
 {
     public class Food : Product
     {
-        public Food(double price, string description, int quantity) : base(price, description, quantity)
+        public Food(decimal price, string description, int quantity) : base(price, description, quantity)
         {
 
         }
