+++ MAM_Assessment/Madical.cs
@@ -3,7 +3,7 @@ namespace MAM_Assessment
 {
     public class Madical : Product
     {
-        public Madical(double price, string description, int quantity) : base(price, description, quantity)
+        public Madical(decimal price, string description, int quantity) : base(price, description, quantity)
         {
 
         }
